pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = "An extensible todo list manager"
authors = ["Yubo Cao <cao2006721@gmail.com>"]
readme = "README.md"
packages = [{include = "todo"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
appdirs = "^1.4.4"
matplotlib = "^3.6.0"
numpy = "^1.23.3"
PyYAML = "^6.0"
lxml = "^4.9.1"
aiohttp = "^3.8.3"
async-property = "^0.2.1"
PySide6 = "^6.3.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.16.0"
black = {version = "^22.8.0", allow-prereleases = true}
pytest = "^7.1.3"
pytest-asyncio = "^0.19.0"
pytest-timeout = "^2.1.0"
mypy = "^0.981"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
